==> 01)  FB_XS_ITA_consol (Optimized).ipynb <==
Import packages

# special IPython command to prepare the notebook for matplotlib
%matplotlib inline

import numpy as np
import pandas as pd
import dask.dataframe as dd
import scipy.stats as stats
import matplotlib.pyplot as plt
import sklearn
import statsmodels.api as sm
from scipy.stats import norm
from sklearn.preprocessing import StandardScaler
from scipy import stats
import warnings
warnings.filterwarnings('ignore')

import seaborn as sns
sns.set_style("whitegrid")
sns.set_context("poster")
color = sns.color_palette()

# special matplotlib argument for improved plots
from matplotlib import rcParams

pd.set_option("display.max_rows",999)
pd.set_option("display.max_columns", 999)

Read in FBXSITA final mailfile

mb_pbfbxs_ita_0518 = pd.read_csv('E:/Pallavi/2018/rev1/rev2/mailfile/mailfiles/561) 117487A_RMC_PBFBXS5.0_18ITA_Final_Mailfile.TXT', 
                                 sep='\t', na_values='.', encoding="ISO-8859-1", error_bad_lines=False, low_memory=False, 
                                parse_dates = ["DOB", "Drop_Date", "Closed_Date", "POffDate"])
mb_pbfbxs_ita_0518.columns = mb_pbfbxs_ita_0518.columns.str.strip().str.lower().str.replace(' ', '_') 

mb_pbfbxs_ita_0518.head()

mb_pbfbxs_ita_0518['zip5'] = mb_pbfbxs_ita_0518['zip5'].astype(int)
mb_pbfbxs_ita_0518['zip5'] = mb_pbfbxs_ita_0518['zip5'].astype(str)
mb_pbfbxs_ita_0518['zip4'] = mb_pbfbxs_ita_0518['zip4'].astype(int)
mb_pbfbxs_ita_0518['zip4'] = mb_pbfbxs_ita_0518['zip4'].astype(str)
mb_pbfbxs_ita_0518['qseqnum'] = mb_pbfbxs_ita_0518['qseqnum'].astype(int)
mb_pbfbxs_ita_0518['qseqnum'] = mb_pbfbxs_ita_0518['qseqnum'].astype(str)
mb_pbfbxs_ita_0518['branchnumber'] = mb_pbfbxs_ita_0518['branchnumber'].astype(int)
mb_pbfbxs_ita_0518['branchnumber'] = mb_pbfbxs_ita_0518['branchnumber'].astype(str)
mb_pbfbxs_ita_0518['distance_to_branch'] = pd.to_numeric(mb_pbfbxs_ita_0518['distance_to_branch'], errors='coerce')
mb_pbfbxs_ita_0518['amt_given1'] = mb_pbfbxs_ita_0518['amt_given1'].str.replace(',', '')
mb_pbfbxs_ita_0518['amt_given1'] = mb_pbfbxs_ita_0518['amt_given1'].str.replace('$', '')
mb_pbfbxs_ita_0518['amt_given1'] = pd.to_numeric(mb_pbfbxs_ita_0518['amt_given1'], errors='coerce')
mb_pbfbxs_ita_0518['up_to_offer'] = mb_pbfbxs_ita_0518['up_to_offer'].str.replace(',', '')
mb_pbfbxs_ita_0518['up_to_offer'] = mb_pbfbxs_ita_0518['up_to_offer'].str.replace('$', '')
mb_pbfbxs_ita_0518['up_to_offer'] = pd.to_numeric(mb_pbfbxs_ita_0518['up_to_offer'], errors='coerce')
mb_pbfbxs_ita_0518['from_offer_amount'] = pd.to_numeric(mb_pbfbxs_ita_0518['from_offer_amount'], errors='coerce')
mb_pbfbxs_ita_0518['fico'] = pd.to_numeric(mb_pbfbxs_ita_0518['fico'], errors='coerce')

Print dimentions

mb_pbfbxs_ita_0518.shape

Print summary of variables

mb_pbfbxs_ita_0518.describe(include = 'all')

Find unique campaign IDs

Set memory usage parameters to `deep` for accurate measure of memory usage

mb_pbfbxs_ita_0518.info(memory_usage='deep')

Average memory usage per data type

for dtype in ['float', 'int', 'object']:
    selected_dtype = mb_pbfbxs_ita_0518.select_dtypes(include = [dtype])
    mean_usage_b = selected_dtype.memory_usage(deep = True).mean()
    mean_usage_mb = mean_usage_b / 1024 ** 2
    print("Average memory usage for {} columns: {:03.2f} MB".format(dtype,mean_usage_mb))

import numpy as np
int_types = ["uint8", "int8", "int16"]
for it in int_types:
    print(np.iinfo(it))

# We're going to be calculating memory usage a lot,
# so we'll create a function to save us some time!

def mem_usage(pandas_obj):
    if isinstance(pandas_obj, pd.DataFrame):
        usage_b = pandas_obj.memory_usage(deep = True).sum()
    else: # we assume if not a df it's a series
        usage_b = pandas_obj.memory_usage(deep = True)
    usage_mb = usage_b / 1024 ** 2 # convert bytes to megabytes
    return "{:03.2f} MB".format(usage_mb)

mb_pbfbxs_ita_0518_int = mb_pbfbxs_ita_0518.select_dtypes(include = ['int'])
converted_int = mb_pbfbxs_ita_0518_int.apply(pd.to_numeric, downcast = 'unsigned')

print(mem_usage(mb_pbfbxs_ita_0518_int))
print(mem_usage(converted_int))

compare_ints = pd.concat([mb_pbfbxs_ita_0518_int.dtypes, converted_int.dtypes], axis = 1)
compare_ints.columns = ['before', 'after']
compare_ints.apply(pd.Series.value_counts)

mb_pbfbxs_ita_0518_float = mb_pbfbxs_ita_0518.select_dtypes(include=['float'])
converted_float = mb_pbfbxs_ita_0518_float.apply(pd.to_numeric,downcast='float')

print(mem_usage(mb_pbfbxs_ita_0518_float))
print(mem_usage(converted_float))

compare_floats = pd.concat([mb_pbfbxs_ita_0518_float.dtypes,converted_float.dtypes],axis=1)
compare_floats.columns = ['before','after']
compare_floats.apply(pd.Series.value_counts)

optimized_mb_pbfbxs_ita_0518 = mb_pbfbxs_ita_0518.copy()

optimized_mb_pbfbxs_ita_0518[converted_int.columns] = converted_int
optimized_mb_pbfbxs_ita_0518[converted_float.columns] = converted_float

print(mem_usage(mb_pbfbxs_ita_0518))
print(mem_usage(optimized_mb_pbfbxs_ita_0518))

from sys import getsizeof

s1 = 'working out'
s2 = 'memory usage for'
s3 = 'strings in python is fun!'
s4 = 'strings in python is fun!'

for s in [s1, s2, s3, s4]:
    print(getsizeof(s))

obj_series = pd.Series(['working out',
                          'memory usage for',
                          'strings in python is fun!',
                          'strings in python is fun!'])
obj_series.apply(getsizeof)

mb_pbfbxs_ita_0518_obj = mb_pbfbxs_ita_0518.select_dtypes(include=['object']).copy()
mb_pbfbxs_ita_0518_obj.describe()

campaign_id = mb_pbfbxs_ita_0518_obj.campaign_id
print(campaign_id.head())

campaign_id_cat = campaign_id.astype('category')
print(campaign_id_cat.head())

campaign_id_cat.head().cat.codes

print(mem_usage(campaign_id))
print(mem_usage(campaign_id_cat))

converted_obj = pd.DataFrame()

for col in mb_pbfbxs_ita_0518_obj.columns:
    num_unique_values = len(mb_pbfbxs_ita_0518_obj[col].unique())
    num_total_values = len(mb_pbfbxs_ita_0518_obj[col])
    if num_unique_values / num_total_values < 0.5:
        converted_obj.loc[:,col] = mb_pbfbxs_ita_0518_obj[col].astype('category')
    else:
        converted_obj.loc[:,col] = mb_pbfbxs_ita_0518_obj[col]

print(mem_usage(mb_pbfbxs_ita_0518_obj))
print(mem_usage(converted_obj))

compare_obj = pd.concat([mb_pbfbxs_ita_0518_obj.dtypes,converted_obj.dtypes],axis=1)
compare_obj.columns = ['before','after']
compare_obj.apply(pd.Series.value_counts)

optimized_mb_pbfbxs_ita_0518[converted_obj.columns] = converted_obj

mem_usage(optimized_mb_pbfbxs_ita_0518)

drop_date = optimized_mb_pbfbxs_ita_0518.drop_date
print(mem_usage(drop_date))
drop_date.head()

optimized_mb_pbfbxs_ita_0518['drop_date'] = pd.to_datetime(drop_date,format='%Y%m%d')

print(mem_usage(optimized_mb_pbfbxs_ita_0518))
optimized_mb_pbfbxs_ita_0518.drop_date.head()

closed_date = optimized_mb_pbfbxs_ita_0518.closed_date
print(mem_usage(closed_date))
closed_date.head()

optimized_mb_pbfbxs_ita_0518['closed_date'] = pd.to_datetime(closed_date,format='%Y%m%d')

print(mem_usage(optimized_mb_pbfbxs_ita_0518))
optimized_mb_pbfbxs_ita_0518.closed_date.head()

dob = optimized_mb_pbfbxs_ita_0518.dob
print(mem_usage(dob))
dob.head()

optimized_mb_pbfbxs_ita_0518['dob'] = pd.to_datetime(dob,format='%Y%m%d', errors = 'coerce')

print(mem_usage(optimized_mb_pbfbxs_ita_0518))
optimized_mb_pbfbxs_ita_0518.dob.head()

poffdate = optimized_mb_pbfbxs_ita_0518.poffdate
print(mem_usage(poffdate))
poffdate.head()

optimized_mb_pbfbxs_ita_0518['poffdate'] = pd.to_datetime(poffdate,format='%Y%m%d', errors = 'coerce')

print(mem_usage(optimized_mb_pbfbxs_ita_0518))
optimized_mb_pbfbxs_ita_0518.poffdate.head()

optimized_mb_pbfbxs_ita_0518.info(memory_usage='deep')

dtypes = optimized_mb_pbfbxs_ita_0518.drop(['drop_date', 'closed_date', 'dob', 'poffdate'],axis=1).dtypes

dtypes_col = dtypes.index
dtypes_type = [i.name for i in dtypes.values]

column_types = dict(zip(dtypes_col, dtypes_type))

# rather than print all 161 items, we'll
# sample 10 key/value pairs from the dict
# and print it nicely using prettyprint

preview = first2pairs = {key:value for key,value in list(column_types.items())[:10]}
import pprint
pp = pp = pprint.PrettyPrinter(indent=4)
pp.pprint(preview)

optimized_mb_pbfbxs_ita_0518.to_csv(r'C:\Users\cepps\RMC\Mail Files\Optimized_Mail_Files\fbxs_ita\optimized_mb_pbfbxs_ita_0518.txt', 
                                    sep = '\t', mode='a')

optimized_mb_pbfbxs_ita_0518 = pd.read_csv('C:/Users/cepps/RMC/Mail Files/Optimized_Mail_Files/fbxs_ita/optimized_mb_pbfbxs_ita_0518.txt', 
                                           dtype = column_types, sep='\t', na_values='.', encoding="ISO-8859-1", error_bad_lines=False, 
                                           low_memory=False, parse_dates = ["dob", "drop_date", "closed_date", "poffdate"], 
                                           infer_datetime_format = True)
optimized_mb_pbfbxs_ita_0518.columns = optimized_mb_pbfbxs_ita_0518.columns.str.strip().str.lower().str.replace(' ', '_') 
print(mem_usage(optimized_mb_pbfbxs_ita_0518))
optimized_mb_pbfbxs_ita_0518.head()

is_ita = read_and_optimized['campaign_id'] == "FBITA5.0_2018"
mb_fb_ita_0518 = read_and_optimized[is_ita]

mb_fb_ita_0518.shape

mb_pbfbxs_ita_0518.campaign_id.unique()
