==> 01)  FB_XS_ITA_consol.ipynb <==
Load libraries

Set parameters to extend java's memory usage to keep xlsx scripts from crashing

# set parameters to extend java's memory usage
# to keep xlsx scripts from crashing
options(java.parameters = "-Xmx8000m")

library(sqldf)

Read in FBXSITA final mailfile

mb_pbfbxs_ita <- read.table("//rmc.local/dfsroot/Marketing/SourceData/Mail files For Data Warehouse/581) 118645A_RMC_PBFBXS8.0_18ITA_Setupheader.TXT", 
                                 fill=TRUE, header=TRUE, sep="\t", quote="")

Print dimentions

dim(mb_pbfbxs_ita)

Print summary of variables

str(mb_pbfbxs_ita)
summary(mb_pbfbxs_ita)

Store `ï..campaign_id` as `campaign_id`

mb_pbfbxs_ita$campaign_id <- mb_pbfbxs_ita$ï..campaign_id

Find unique campaign IDs

unique(mb_pbfbxs_ita$Campaign_id)

Subset the FBITA campaign

mb_fb_ita_1 <- sqldf(
  "
  SELECT * 
  FROM mb_pbfbxs_ita 
  WHERE campaign_id IN ('FBITA5.0_2018')
  "
  )

Print dimensions

dim(mb_fb_ita_1)

Print summary of variables

str(mb_fb_ita_1)
summary(mb_fb_ita_1)

Print column names

colnames(mb_fb_ita_1)

Print `Drop_Date` variable type

typeof(mb_fb_ita_1$Drop_Date)

Convert `Drop_Date` from integer to date type variable

mb_fb_ita_1$Drop_Date <- as.Date("2018-05-07", format = "%Y-%m-%d", origin = "1970-01-01")

Save FBITA R data object

save(mb_fb_ita_1,
     file = "C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_fb_ita_0518.RData")

Load master FBITA R data object

load("E:/Pallavi/2018/rev1/rev2/mailfile/mailfiles/mb_fb_ita.RData")

load("C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_fb_ita.RData")

unique(mb_fb_ita$camp_type1)

Convert FBITA master set variable, `Drop_Date`, from double to date type

mb_fb_ita$Drop_Date <- as.Date(mb_fb_ita$Drop_Date,
                               format = "%Y-%m-%d",
                               origin = "1970-01-01")

write.table(mb_fb_ita, "C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_fb_ita_ALL.txt", sep="\t", col.names = TRUE)

Subset and join the master FBITA and current FBITA files

mb_fb_ita <- sqldf(
   "
   select mon_ind, 
          camp_type, 
          camp_type1, 
          ss7, 
          Drop_Date, 
          NAME, 
          STREET,
          ZIP5, 
          fico, 
          apr, 
          offer_amount, 
          state, 
          BranchNumber 
   from mb_fb_ita
   
   union all

   select 104 as mon_ind, 
          '05.FB_ITA' as camp_type,
          'FB_ITA_0818__Aug-18' as camp_type1, 
          SSN as ss7, 
          Drop_Date,
          firstName||' '||lastName as NAME, 
          addressOne as STREET, 
          zip as ZIP5, 
          0 as fico, 
          0 as apr, 
          0 as offer_amount, 
          state, 
          Branch as BranchNumber 
   from mb_fb_ita_1
   "
   )

mb_fb_ita <- sqldf(
   "
   select mon_ind, 
          camp_type, 
          camp_type1, 
          ss7, 
          Drop_Date, 
          NAME, 
          STREET,
          ZIP5, 
          fico, 
          apr, 
          offer_amount, 
          state, 
          BranchNumber 
   from mb_fb_ita
   
   union all

   select 101 as mon_ind, 
          '05.FB_ITA' as camp_type,
          'FB_ITA_0518_May-18' as camp_type1, 
          ss7, 
          Drop_Date,
          name1 as NAME, 
          STREET, 
          ZIP5, 
          0 as fico, 
          0 as apr, 
          0 as offer_amount, 
          state, 
          BranchNumber 
   from mb_fb_ita_1
   "
   )

Save condensed master FBITA R data object

save(mb_fb_ita, 
     file = "C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/new_consol/mb_fb_ita.RData")

Subset auto and retail cross sell campaigns as one cross sell data set

mb_xs_ita_1 <- sqldf(
   "
   select * 
   from mb_pbfbxs_ita 
   where campaign_id in ('RetailXSITA5.0_2018', 'AutoXSITA5.0_2018')
   "
   )

Print dimensions

dim(mb_xs_ita_1)

Print summary of variables

str(mb_xs_ita_1)
summary(mb_xs_ita_1)

Convert `Drop_Date` from integer to date type

mb_xs_ita_1$Drop_Date <- as.Date("2018-05-07", format="%Y-%m-%d", origin="1970-01-01")

typeof(mb_xs_ita_1$Drop_Date)

Save XSITA dataset as R data oblect

save(mb_xs_ita_1, 
     file = "C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_xs_ita_0518.RData")

Load XSELLITA master dataset

load("E:/Pallavi/2018/rev1/rev2/mailfile/mailfiles/mb_xsell_ita.RData")

load("C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_xsell_ita.RData")

unique(mb_xsell_ita$camp_type1)

write.csv(mb_xsell_ita, file = "C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_xsell_ita_ALL.csv")

write.table(mb_xsell_ita, "C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_xsell_ita_ALL.txt", sep="\t", col.names = TRUE)

str(mb_xsell_ita)
summary(mb_xsell_ita)

Convert `Drop_Date` from integer to date type

mb_xsell_ita$Drop_Date <- as.Date(mb_xsell_ita$Drop_Date,
                                  format = "%Y-%m-%d",
                                  origin = "1970-01-01")

Subset and join the master XSELLITA and current XSELLITA files

mb_xsell_ita <- sqldf(
   "
   select mon_ind, 
          camp_type, 
          camp_type1, 
          ss7, 
          Drop_Date, 
          NAME, 
          STREET,
          ZIP5, 
          fico, 
          apr, 
          offer_amount, 
          state, 
          BranchNumber 
   from mb_xsell_ita
   
   union all

   select 104 as mon_ind, 
          '07.XSELL_ITA' as camp_type,
          'XSELL_ITA_0818' as camp_type1, 
          SSN as ss7, 
          Drop_Date,
          firstName||' '||lastName as NAME, 
          addressOne as STREET, 
          zip as ZIP5, 
          0 as fico, 
          0 as apr, 
          0 as offer_amount, 
          state, 
          Branch as BranchNumber 
   from mb_xs_ita_1
   "
   )

mb_xsell_ita <- sqldf(
   "
   select mon_ind, 
          camp_type, 
          camp_type1, 
          ss7, 
          Drop_Date, 
          NAME, 
          STREET,
          ZIP5, 
          fico, 
          apr, 
          offer_amount, 
          state, 
          BranchNumber 
   from mb_xsell_ita

   union all

   select 101 as mon_ind, 
          '07.XSELL_ITA' as camp_type, 
          'XSELL_ITA_0518' as camp_type1, 
          ss7, 
          Drop_Date,
          name1 as NAME, 
          STREET, 
          ZIP5 as ZIP5, 
          fico, 
          0 as apr,
          0 as offer_amount, 
          state, 
          BranchNumber 
   from mb_xs_ita_1
   "
   )

Save condensed master XSELLITA dataset as R data object

save(mb_xsell_ita, 
     file = "C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/new_consol/mb_xsell_ita.RData")
