==> 01) FB_XS_ITA_python.ipynb <==
# special IPython command to prepare the notebook for matplotlib
%matplotlib inline

import pandas as pd
import pandasql as pdsql
import warnings
import numpy as np
warnings.filterwarnings('ignore')
import seaborn as sns
sns.set_style("whitegrid")
sns.set_context("poster")
color = sns.color_palette()
pd.set_option("display.max_rows",999)
pd.set_option("display.max_columns", 999)

df = pd.read_csv(r'\\rmc.local\dfsroot\Marketing\SourceData\Mail files For Data Warehouse\581) 118645A_RMC_PBFBXS8.0_18ITA_Final_Mailfile.TXT', 
                 sep='\t', na_values='.', encoding="ISO-8859-1", error_bad_lines=False, low_memory=False)
df.columns = df.columns.str.strip().str.lower().str.replace(' ', '_') 
df['drop_date'] = pd.to_datetime(df['drop_date'],infer_datetime_format=True)
df['drop_date'] = df['drop_date'].astype(str)
df.head()

df = pd.read_csv(r'C:\Users\cepps\RMC\Mail Files\R_data_objects\fbxs_ita\mb_fb_ita_ALL.txt', 
                 sep='\t', na_values='.', encoding="ISO-8859-1", error_bad_lines=False, low_memory=False)
df.columns = df.columns.str.strip().str.lower().str.replace(' ', '_') 
df['drop_date'] = pd.to_datetime(df['drop_date'],infer_datetime_format=True)
df['drop_date'] = df['drop_date'].astype(str)
df.head()

df.shape

np.round(df.describe(include = 'all'), 2).T

df.campaign_id.unique()

pysql = lambda q: pdsql.sqldf(q, globals())
str1 = """
       SELECT * 
       FROM df 
       WHERE campaign_id IN ('FBITA1.0_2018');
       """

mb_fb_ita = pysql(str1)
mb_fb_ita.head()

mb_fb_ita.shape

np.round(mb_fb_ita.describe(include = 'all'), 2).T

list(mb_fb_ita)

df = pd.read_csv(r'C:\Users\cepps\RMC\Mail Files\R_data_objects\fbxs_ita\mb_xsell_ita_ALL.txt', 
                 sep='\t', na_values='.', encoding="ISO-8859-1", error_bad_lines=False, low_memory=False)
df.columns = df.columns.str.strip().str.lower().str.replace(' ', '_') 
df['drop_date'] = pd.to_datetime(df['drop_date'],infer_datetime_format=True)
df['drop_date'] = df['drop_date'].astype(str)
df.head()

df.drop_date.unique()

df.groupby(["camp_type", "camp_type1", "drop_date"]).size().reset_index(name="count")
