==> 03)  NBPQLL_consol.ipynb <==
# set parameters to extend java's memory usage
# to keep xlsx scripts from crashing
options(java.parameters = "-Xmx8000m")

library(sqldf)

df <- read.table("//rmc.local/dfsroot/Marketing/SourceData/Mail files For Data Warehouse/550) 116960A_RMC_NB3.1_0418PQ_Final_Mailfile.txt",
                 fill = TRUE, header = TRUE, sep = "\t", quote = "")

nrow(df)

df <- sqldf(
  "
  select *, 
    case when length(SS7) = 5 then 'A00'|| SS7 
         when length(SS7) = 6 then 'A0'|| SS7 
          else 'A' || SS7 end as SSN7 
  from df a
  "
  )

unique(df$Drop_Date)

df$Drop_Date <- '04/02/2018'

df$Drop_Date <- as.Date(df$Drop_Date, format = "%m/%d/%Y")

str(df$Drop_Date)

unique(df$Drop_Date)

save(df, file = "C:/Users/cepps/RMC/Mail Files/R_data_objects/nb_pq_ll/mb_nbpqll_0550.RData")

save(df, file = "C:/Users/cepps/RMC/Mail Files/R_data_objects/nb_pq_ll/mb_nbpqll.rda")

Append to consolidated NBPQLL mailfile

load("C:/Users/cepps/RMC/Mail Files/R_data_objects/nb_pq_ll/mb_nbpqll.rda")

head(mb_nbpqll)

nrow(df)

mb_nbpqll <- sqldf(
  "
  select  mon_ind, camp_type, camp_type1, ss7, Drop_Date, NAME, STREET,
          ZIP5, fico, apr, offer_amount, state, BranchNumber 
  from mb_nbpqll
  
  union all

  select 101 as mon_ind, '01.NBPQLL' as camp_type,
         'NBPQLL_0518__May-18' as camp_type1, ss7, Drop_Date,
         name as NAME, STREET, ZIP5 as ZIP5, fico, APR as apr,
         OFFER_AMOUNT as offer_amount, state, BranchNumber 
  from df
  "
  )

nrow(mb_nbpqll)

save(mb_nbpqll,
     file = "C:/Users/cepps/RMC/Mail Files/R_data_objects/nb_pq_ll/mb_nbpqll.RData")
