==> 04)  NBCC_consol.ipynb <==
library(dplyr)
library(sqldf)

# Import new mailfile
df <- read.table("E:/Pallavi/2018/rev1/rev2/mailfile/mailfiles/567) 118144A_RMC_NB_5.0_Final_mailfile.txt", 
                 fill=TRUE, header=TRUE, sep="\t", quote="")
nrow(df)

# Check Drop_Date
unique(df$Drop_Date)

str(df$Drop_Date)

# CHANGE DATE and date format if required
df$Drop_Date <- as.Date("2018-06-19",format="%Y-%m-%d",origin="1970-01-01")
str(df$Drop_Date)

unique(df$Drop_Date)

# Load latest consolidated mailfile
load("C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_fb_ita.RData")

write.table(mb_fb_ita, "C:/Users/cepps/RMC/Mail Files/R_data_objects/fbxs_ita/mb_fb_ita_ALL.txt", sep="\t")

mb_nbcc$Drop_Date <- as.Date(mb_nbcc$Drop_Date,format="%Y-%m-%d",origin="1970-01-01")
nrow(mb_nbcc)

unique(mb_fb_ita$Drop_Date)

df <- sqldf(
    "
    select *,
        case when length(SS7) = 5 then '00'||SS7 
             when length(SS7)=6 then '0'||SS7 
        else SS7 end as SSN7 
    from df a
    "
)

df <- sqldf(
  "
  select *, 
    case when length(SS7) = 5 then 'A00'|| SS7 
         when length(SS7) = 6 then 'A0'|| SS7 
          else 'A' || SS7 end as SSN7 
  from df a
  "
  )

df$CheckAmount <- gsub("$","",df$CheckAmount,fixed=TRUE)
df$CheckAmount <- gsub(",","",df$CheckAmount,fixed=TRUE)
df$CheckAmount <- as.integer(df$CheckAmount)

mb_nbcc<- sqldf("select mon_ind, camp_type, camp_type1,
                ss7,Drop_Date, NAME, STREET, ZIP5 as ZIP5, fico, apr, 
                offer_amount, state, BranchNumber 
                from mb_nbcc
                
                union all
                
                select '102' as mon_ind,'02.NBCC' as camp_type,'NBCC_0618__Jun-18_Drop-1' as camp_type1,
                SS7 as ss7, Drop_Date as Drop_Date, FName as NAME, STREET, ZIP5, fico, Apr1 as apr,
                CheckAmount as offer_amount, state, BranchNumber 
                from df
                ")

unique(mb_nbcc$camp_type1)

nrow(mb_nbcc)

write.table(mydata, "c:/mydata.txt", sep="\t")

save(mb_nbcc_0618, file="E:/Pallavi/2018/rev1/rev2/mailfile/mailfiles/mb_nbcc_0618.RData")

save(mb_nbcc,file="E:/Pallavi/2018/rev1/rev2/mailfile/mailfiles/new_consol/mb_nbcc.RData")
